requirements.txt
pre-commit
ruff
isort
mypy
Pillow
